+++ Sketch/libraries/StepperLib/src/Analog8IOControlSmooth.h
@@ -96,7 +96,7 @@ public:
 
 	void Poll()
 	{
-		unsigned long milli;
+		uint32_t milli;
 		if (_currentlevel != _iolevel && (milli = millis()) >= _nexttime)
 		{
 			_nexttime = milli + _delayMs;
@@ -131,7 +131,7 @@ private:
 		CHAL::analogWrite8(PWMPIN, lvl);
 	}
 
-	unsigned long _nexttime; // time to modify level
+	uint32_t _nexttime; // time to modify level
 #ifndef REDUCED_SIZE
 	uint8_t _level; // value if "enabled", On/Off will switch between 0..level
 #endif
